renamed lib and updated notebook
